requirements.txt
streamlit==1.29.0
google-generativeai==0.3.2
python-docx==1.1.0
requests==2.31.0
python-dotenv==1.0.0
